New Payment Integration
This update creates a client response server response system to mediate a purchase between a user a the server.

A Stripe message handler creates a message when the purchase is clicked. Sending a message to the server requesting a Json message. The message contains the list of items and the token id of the client. Then remove all items in the cart if the message was sent.

The server cross references the items and sends the total to stripe charges, returning a message to the client.