new navigation
Changing the styling to fit the new theme. The theme is going to be a simple white back ground store with a modern feed and modern text.

The style of the navigation and styling for the header is very minimalistic. Logo with a brand name and slogan with links to other pages and socials.

More implementation for smaller screen sizes must be made.

Implementation of a mobile navigation is not done yet. And implementation of a footer is also not done yet.

Restyling the content will take some time which will be the main blocker going foreword.

Removing old styles from the scraped project.