Stripe Integration Update
adding items to the stripe products store to reference during the checkout process.

post method - "purchase"
pings the server with the list of items in the cart. The server will cross examine the items with the ones on file (checks product number). Product numbers are set on the dashboard, make sure to copy any products listed on the stripe dashboard on file for cross referencing before sending it up to stripe for payment details. Then changes the page to the payment page.

Server will get the cart items and cross reference against server side products, then it will format the data and then create a checkout session.  The session url is sent back to the server. An alert is sent from the server then the page is switched.

note: do not use store-init
-only meant for testing purposes
-create all products from the stripe API
-do not try to create a product from code