Storing Cart on Client Side Local Storage
When the page changes the information in the cart is lost.

To ensure better conversion, the client's cart will be stored locally using the class Cart.

data fields: cart, which is an array
methods:
add - adds parameters to an item object and stores it to the cart array.
getTotal - returns the net cost of all items
saveCartLocally stores the cart as a JSON under the key 'Arte-Cart'

The cart can be retrieved using the same file with a external function getCartFromLocal which returns the Parsed JSON with the cart array.