Loading Cart from Session Storage
In the ready function, if the cart array is found by getCartFromLocal in the cart module, then check if the array was passed.

For each element in the array add the item to the cart.

Then update the total using the function in this file.

AddItemCart is modified to implement a quantity input for multiples of items from the cart.  Stores either quantity or  1 in quantityMax, then renders the value as ejs. Then checks whether the quantity was input, thus a not a new item, to ensure that the user is spammed with alerts when refreshing the page.

Still to do:
fix the imgSrc issue, as images are not loading in properly.