Creating a Successful Purchase Page
Redirect back to the website after purchasing from stripe and get rid of cart on local storage in this page.

Otherwise keep the cart in session storage and redirect back to the store page.

The redirects from this page are to home and contacts. More styling and functionality must be implemented for the success information.