Added a footer
Created a footer with ejs.

Footer is a form that sends the user email to the server for newsletter implementation.

Added the links to styles and the scripts to the appropriate files as well. Might have to template a header to make importing to all files easier.

Got rid of the old styling for the page and the html.

Added google fonts to the styles.css and used inline styles on the footer to wait till rendering to use the fonts.