Stripe Products Update
Created products id in the stripe dashboard for all products and assigned the new ids.

Please do not use the STORE-INIT path to initialize unauthorized products. This part of the server must be kept of test purposes only. Any changes on products should be made from the dashboard and manually transferred to the server.