Functional Contact Page
Created the outline for the contact page and implemented the form that emails the client when a user submits a message. This was done using the smtp.js library that allows smtp communication. Must go onto the elastic email dashboard and create an account. Verify, do the 2 step authentication, and create a SMTP authentication token. Now emails from the smtp account with the details entered by the user. This cannot be directly sent from the user because they did not configure a smtp server. So the message and the email name is sent in the body text.